qtdemux: fix allocation explosion with stsd entries
Validate that the actual size of stsd box in a stream is same as
the stsd_entries_length property which is used by qtdemux to
allocate memory. This fix prevents a malicious file with arbitrary
length of stsd entries from allocating more memory than required.
This change is derived from mainstream gst v1.16 qtdemux plugin.
https://git.codelinaro.org/clo/ype/external/gstreamer.org/gstreamer/gst-plugins-good/-/commit/310dec75cd0b630a765ee93c82a473e17fb5dd94

Change-Id: Ib1bebe7c4df22b3c089c138ca7b753ed058da9cb