Revert "meta-qti-gst: waylandsink: Remove call to gbm_bo_destroy for dup fd"
This reverts commit f32256b32cb4e84935a4e234fc5695edfda3a616.

Reason for revert:
As per confirmation from display team, gbm_bo_destroy()
call after every gbm_bo_import() is required for proper handling of
buffer objects. Previously, call to this API was removed as display
team had suggested this fix as more than one calls to gbm_bo_import()
API were failing for the same FD. Upon further analysis it is was
confirmed that call to gbm_bo_destroy() is necessary and this is not
the root cause of the issue.

Change-Id: I4efae96914b828817d4bef0cc77b488667d05291