meta-qti-gst: v4l2: Remove call to gbm_bo_get_fd() API
Issue: When a GBM buffer object is created, gbm_bo_get_fd() API is
used to get the fd of the buffer object. This API creates and returns
a dup'd FD. There is a limitation from GBM side that, gbm_bo_import()/
gbm_bo_destroy() cannot be called more that once on a dup'd FD.
This is causing display related issues in certain usecases(e.g. seek).

Fix: Directly access the original FD from buffer object once the buffer
object is created as the handling for original FD is proper on GBM side.

Change-Id: Idd1e743c1e7acd63c48d0ab6f1e4af377e96fe24