meta-qti-gst: v4l2: Update caps feature in decoder negotiate
Issue: In decoder negotiate, when gstinfo object is created from
caps, any caps feature related info is not retained. So the caps
feature info (memory GBM feature) is not present in the final
caps used for negotitation with downstream element.

Fix: Check if the memory feature is present in the peer compatible
fixated caps and add the feature to caps created from gstinfo object
after set_format (fixated caps might be modified in set_format).

Change-Id: I758d4a99512d9b363c81628777826843dfa94107