meta-qti-gst: adding license files
During do_rootfs of the image, we are getting warnings stating,
"license is not found in license collected in the recipe".

This is because of the lic checksum for those licenses not added.
license gets included in the license directory, where
during do_rootfs it searches for the licenses.
Licenses gets added only when it is mentioned in LIC_FILES_CHKSUM.

Change-Id: I66a8ea294768ba229691f7d316a2d37a0aa5cd27